+++ mockbird-server/src/main/java/com/mockbird/service/ApiInterfaceService.java
@@ -1,0 +1,7 @@
+package com.mockbird.service;
+
+import com.baomidou.mybatisplus.extension.service.IService;
+import com.mockbird.entity.ApiInterface;
+
+public interface ApiInterfaceService extends IService<ApiInterface> {
+}