+++ mockbird-server/src/main/java/com/mockbird/service/MockRuleService.java
@@ -1,0 +1,7 @@
+package com.mockbird.service;
+
+import com.baomidou.mybatisplus.extension.service.IService;
+import com.mockbird.entity.MockRule;
+
+public interface MockRuleService extends IService<MockRule> {
+}