+++ mockbird-server/src/main/java/com/mockbird/service/RequestLogService.java
@@ -1,0 +1,7 @@
+package com.mockbird.service;
+
+import com.baomidou.mybatisplus.extension.service.IService;
+import com.mockbird.entity.RequestLog;
+
+public interface RequestLogService extends IService<RequestLog> {
+}