+++ mockbird-server/src/main/java/com/mockbird/service/impl/ProjectServiceImpl.java
@@ -1,0 +1,11 @@
+package com.mockbird.service.impl;
+
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.mockbird.entity.Project;
+import com.mockbird.mapper.ProjectMapper;
+import com.mockbird.service.ProjectService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProjectServiceImpl extends ServiceImpl<ProjectMapper, Project> implements ProjectService {
+}